Address #340 review: docstring style, TDD, x64+AOT guard, period dtype
Docstring/comment cleanup (drop counterfactuals and history rehashes per
AGENTS.md "Describe state, not history"):

- discrete.py to_jax: rephrase "would otherwise produce" as positive
  constraint
- initial_conditions.py: replace "Without this..." with the invariant the
  cast enforces
- compile.py module docstring: drop "Mirrors the pattern in solve_brute"
  cross-reference
- compile.py argmax-over-active-periods comment: drop aca-model-specific
  forced-canwork / pre-FRA example
- compile.py _build_next_state_args: drop the now-stale comment about
  Python-int period (all sites use jnp.int32)

Bullet-list dispatch cases (per AGENTS.md "Bulleted lists for enumerated
cases"):

- model.py n_subjects field: 4 dispatch cases as bullets, plus an explicit
  param-shape stability contract for MSM-style use
- model.py _resolve_simulate_internal_regimes: 3 cases as bullets

TDD/test style (per AGENTS.md "Testing"):

- test_simulate_aot.py: split test_n_subjects_none_keeps_lazy_behavior
  into two single-assertion tests
- test_simulate_aot.py: rename test_simulate_caches_recompiled_size_no_
  second_warning -> test_simulate_warns_only_once_per_mismatching_size
  (the implementation only adds to _warned_n_subjects, not the cache)
- test_simulate_aot.py: split first/second-call compile counting into
  test_simulate_first_matching_call_populates_aot_cache and
  test_simulate_second_matching_call_does_not_invoke_compile, dropping
  the weak `n_first > 0` smoke check
- test_simulate_aot.py: docstrings on the validation tests, return-type
  annotations on counting_compile and _build_initial_conditions
- test_int_dtype_invariants.py: per-test docstrings
- compile.py _get_regime_V_shapes: docstring matching the file's other
  private helpers
- model.py: class-body annotations and field docstrings for
  _simulate_compile_cache, _warned_n_subjects, _simulate_compile_lock

x64 + AOT fail-fast (issue 19):

- model_processing._fail_if_x64_with_aot raises
  ModelInitializationError when n_subjects is set under
  jax_enable_x64=True; the AOT path pins integer dtypes to int32, x64
  promotes to int64, so the cached signature would not match runtime
- AOT tests get an autouse fixture that disables x64 for the duration

period dtype consistency (issue 20):

- _build_next_state_args / _build_crtp_args: lower with
  period=jnp.int32(0) to match the argmax path
- transitions.calculate_next_states / calculate_next_regime_membership:
  pass period=jnp.int32(period) at runtime so the runtime call matches
  the AOT abstract signature

Concurrency (issue 18):

- Model gains a threading.Lock guarding check-then-set on
  _simulate_compile_cache and _warned_n_subjects
- __getstate__ / __setstate__ exclude the lock and the per-process AOT
  cache from pickling (compiled programs can't survive process boundaries
  anyway)

Dedup-key invariant (issue 17):

- _collect_unique_simulate_functions: comment documents the dedup
  contract — pylcm's process_regimes ships per-regime callables for
  next_state and crtp, so identity-based dedup is collision-free
- new test_simulate_functions_use_per_regime_callables in
  test_regime_processing.py pins the invariant against future
  regression

+++ src/lcm/simulation/transitions.py
@@ -126,7 +126,7 @@ def calculate_next_states(
         **state_action_space.states,
         **optimal_actions,
         **stochastic_variables_keys,
-        period=period,
+        period=jnp.int32(period),
         age=age,
         **regime_params,
     )
@@ -187,7 +187,7 @@ def calculate_next_regime_membership(
         internal_regime.simulate_functions.compute_regime_transition_probs(  # ty: ignore[call-non-callable]
             **state_action_space.states,
             **optimal_actions,
-            period=period,
+            period=jnp.int32(period),
             age=age,
             **regime_params,
         )
